Finalize PS3 port and resolve legacy compiler incompatibilities
- Fixed ambiguous std::pow call in Script.cpp for legacy toolchains.
- Replaced binary literals with hex in Drawing.cpp.
- Guarded all SDL-specific members and calls in Drawing.cpp.
- Moved class member initializations to the RetroEngine constructor.
- Replaced nullptr with NULL.
- Defined OpenGL OES aliases and GL_COLOR_ATTACHMENT0_OES for PSGL.
- Resolved various sign conversion and implicit conversion warnings.
- Fixed unbalanced preprocessor blocks and missing braces.

+++ RSDKv3/Script.cpp
@@ -1495,7 +1495,7 @@ void AppendIntegerToString(char *text, int value)
 
     v = 0;
     for (int i = cnt - 1; i >= 0; --i) {
-        v = value / pow(10, i);
+        v = value / (int)pow(10.0, i);
         v %= 10;
 
         int strValue = v + '0';
